perf: switch to using map[Point]int
refactor: create additional functions / methods

+++ 2019/go/utils/utils.go
@@ -18,7 +18,7 @@ func GetFile(path string) *os.File {
 }
 
 // Abs returns the absolute value of an integer
-func Abs(x int64) int64 {
+func Abs(x int) int {
 	if x < 0 {
 		return -x
 	}
